[Add]予約の取消不可になるバリデーションを作成

+++ app/controllers/reservations_controller.rb
@@ -77,6 +77,10 @@ class ReservationsController < ApplicationController
     @user = User.find(@reservation.user_id)
     @teacher = Teacher.find(@reservation.teacher_id)
     @start_time = @reservation.start_time
+    message = Reservation.check_delete(@start_time)
+    if !!message
+      redirect_to user_path(current_user.id), flash: { alert: message } and return #and returnはリダイレクトと同時にメソッドの実行を終了
+    end 
     if @reservation.destroy
       TeacherMailer.with(user: @user, teacher: @teacher, start_time: @start_time).reservation_delete_email.deliver_later #withメソッドでアクションメーラーのメソッドに必要な情報を渡す。
       flash[:success] = "予約を削除しました。"
